requirements.txt
openai>=1.0.0
python-dotenv>=0.19.0
tenacity>=8.0.0
# logging is part of Python's standard library, so no need to install it separately
